+++ cex.h
@@ -3015,6 +3015,7 @@ void _cex__codegen_indent(_cex__codegen_s* cg);
 #    endif // #ifndef cexy$cex_self_cc
 
 #    ifndef cexy$cc
+/// Default compiler for building tests/apps (by default inferred from ./cex tool compiler)
 #        define cexy$cc cexy$cex_self_cc
 #    endif // #ifndef cexy$cc
 
@@ -3086,13 +3087,12 @@ void _cex__codegen_indent(_cex__codegen_s* cg);
 #    endif
 
 #    ifndef cexy$ld_args
-/// Linker flags (e.g. -L./lib/path/ -lmylib -lm) (may be overridden by user)
+/// Linker flags (e.g. -L./lib/path/ -lmylib -lm) (may be overridden)
 #        define cexy$ld_args
 #    endif
 
 #    ifndef cexy$debug_cmd
-/// Command for running debugger for cex test/app debug  (may be overridden by
-/// user)
+/// Command for launching debugger for cex test/app debug (may be overridden)
 #        define cexy$debug_cmd "gdb", "-q", "--args"
 #    endif
 
@@ -3132,51 +3132,58 @@ See `cex help str.match` for more information about patter syntax.
 #    endif
 
 
-#    define cexy$cmd_process                                                                       \
+#    define _cexy$cmd_process                                                                       \
         { .name = "process",                                                                       \
           .func = cexy.cmd.process,                                                                \
           .help = "Create CEX namespaces from project source code" }
 
-#    define cexy$cmd_new { .name = "new", .func = cexy.cmd.new, .help = "Create new CEX project" }
+#    define _cexy$cmd_new { .name = "new", .func = cexy.cmd.new, .help = "Create new CEX project" }
 
-#    define cexy$cmd_help                                                                          \
+#    define _cexy$cmd_help                                                                          \
         { .name = "help",                                                                          \
           .func = cexy.cmd.help,                                                                   \
           .help = "Search cex.h and project symbols and extract help" }
 
-#    define cexy$cmd_config                                                                        \
+#    define _cexy$cmd_config                                                                        \
         { .name = "config",                                                                        \
           .func = cexy.cmd.config,                                                                 \
           .help = "Check project and system environment and config" }
 
-#    define cexy$cmd_libfetch                                                                      \
+#    define _cexy$cmd_libfetch                                                                      \
         { .name = "libfetch",                                                                      \
           .func = cexy.cmd.libfetch,                                                               \
           .help = "Get 3rd party source code via git or install CEX libs" }
 
+#    define _cexy$cmd_stats                                                                         \
+        { .name = "stats",                                                                         \
+          .func = cexy.cmd.stats,                                                                  \
+          .help = "Calculate project lines of code and quality stats" }
+
+/// Simple test runner command (test runner, debugger launch, etc)
 #    define cexy$cmd_test                                                                          \
         { .name = "test",                                                                          \
           .func = cexy.cmd.simple_test,                                                            \
           .help = "Generic unit test build/run/debug" }
 
-#    define cexy$cmd_stats                                                                         \
-        { .name = "stats",                                                                         \
-          .func = cexy.cmd.stats,                                                                  \
-          .help = "Calculate project lines of code and quality stats" }
 
+/// Simple app build command (unity build, simple linking, runner, debugger launch, etc)
 #    define cexy$cmd_app                                                                           \
         { .name = "app", .func = cexy.cmd.simple_app, .help = "Generic app build/run/debug" }
 
+/// All built-in commands for ./cex tool
 #    define cexy$cmd_all                                                                           \
-        cexy$cmd_help, cexy$cmd_process, cexy$cmd_new, cexy$cmd_stats, cexy$cmd_config,            \
-            cexy$cmd_libfetch
+        _cexy$cmd_help, _cexy$cmd_process, _cexy$cmd_new, _cexy$cmd_stats, _cexy$cmd_config,            \
+            _cexy$cmd_libfetch
 
+/// Initialize CEX build system (build itself)
 #    define cexy$initialize() cexy.build_self(argc, argv, __FILE__)
-
+/// ./cex --help description
 #    define cexy$description "\nCEX language (cexy$) build and project management system"
+/// ./cex --help usage
 #    define cexy$usage " [-D] [-D<ARG1>] [-D<ARG2>] command [options] [args]"
 
 // clang-format off
+/// ./cex --help epilog
     #define cexy$epilog \
         "\nYou may try to get help for commands as well, try `cex process --help`\n"\
         "Use `cex -DFOO -DBAR config` to set project config flags\n"\
@@ -15032,6 +15039,7 @@ _cexy__colorize_ansi(str_s token, str_s exact_match, char current_char)
         "\033[1;32m", // types
         "\033[1;34m", // func/macro call
         "\033[1;35m", // #macro
+        "\033[33m",   // macro const
     };
     static struct
     {
@@ -15089,15 +15097,20 @@ _cexy__colorize_ansi(str_s token, str_s exact_match, char current_char)
                 }
             }
         }
-        if (current_char == '(') {
-            // looks like function/macro call
-            return types[3];
-        }
+        // looks like function/macro call
+        if (current_char == '(') { return types[3]; }
+
+        // CEX style type suffix
         if (str.slice.ends_with(token, str$s("_s")) || str.slice.ends_with(token, str$s("_e")) ||
             str.slice.ends_with(token, str$s("_c"))) {
             return types[2];
         }
+
+        // #preproc directive
         if (token.buf[0] == '#') { return types[4]; }
+
+        // some$macro constant
+        if (str.slice.index_of(token, str$s("$")) >= 0) { return types[5]; }
     }
 
     return "\033[0m"; // no color, not matced
@@ -15194,21 +15207,38 @@ _cexy__display_full_info(
             bool has_macro = false;
             mem$scope(tmem$, _)
             {
+                hm$(str_s, cex_decl_s*) macros = hm$new(macros, _, .capacity = 64);
                 for$each (it, cex_ns_decls) {
-                    if (it->type == CexTkn__macro_func && str.slice.starts_with(it->name, name) &&
-                        it->name.buf[name.len] == '$') {
+                    if (!(it->type == CexTkn__macro_func || it->type == CexTkn__macro_const)) {
+                        continue;
+                    }
+                    if (str.slice.starts_with(it->name, name) && it->name.buf[name.len] == '$') {
                         if (!has_macro) {
                             io.printf("\n");
                             has_macro = true;
                         }
-                        if (it->docs.buf) {
-
-                            str_s brief_str = _cexy__process_make_brief_docs(it);
-                            if (brief_str.len) { io.printf("/// %S\n", brief_str); }
+                        if (hm$getp(macros, it->name) != NULL) {
+                            if (it->docs.buf) { hm$set(macros, it->name, it); }
+                            continue; // duplicate
                         }
-                        char* l = str.fmt(_, "#define %S(%s)\n", it->name, it->args);
-                        _cexy__colorize_print(str.sstr(l), name);
+                        hm$set(macros, it->name, it);
                     }
+                }
+                // WARNING: sorting of hashmap items is a dead end, hash indexes get invalidated
+                qsort(macros, hm$len(macros), sizeof(*macros), str.slice.qscmp);
+
+                for$each (it, macros) {
+                    if (it.value->docs.buf) {
+                        str_s brief_str = _cexy__process_make_brief_docs(it.value);
+                        if (brief_str.len) { io.printf("/// %S\n", brief_str); }
+                    }
+                    char* l = NULL;
+                    if (it.value->type == CexTkn__macro_func) {
+                        l = str.fmt(_, "#define %S(%s)\n", it.value->name, it.value->args);
+                    } else {
+                        l = str.fmt(_, "#define %S\n", it.value->name);
+                    }
+                    _cexy__colorize_print(str.sstr(l), name);
                 }
             }
             io.printf("\n\n");
@@ -15581,10 +15611,7 @@ cexy__cmd__config(int argc, char** argv, void* user_ctx)
                         str.join((const char**)args, arr$len(args), " ", _)
                     );
                 } else {
-                    io.printf(
-                        "* pkg-config (libs test)    %s[%s]\n",
-                        "ERROR", err
-                    );
+                    io.printf("* pkg-config (libs test)    %s[%s]\n", "ERROR", err);
                     result = "Missing Libs";
                 }
             }